+++ ApiKarbord/Models/Public/Web_Dftr.cs
@@ -105,15 +105,15 @@
         [StringLength(100)]
         public string AccName { get; set; }
 
-        public int AccCode1 { get; set; }
+        public long AccCode1 { get; set; }
 
-        public int AccCode2 { get; set; }
+        public long AccCode2 { get; set; }
 
-        public int AccCode3 { get; set; }
+        public long AccCode3 { get; set; }
 
-        public int AccCode4 { get; set; }
+        public long AccCode4 { get; set; }
 
-        public int AccCode5 { get; set; }
+        public long AccCode5 { get; set; }
 
         [StringLength(50)]
         public string MkzCode { get; set; }
@@ -121,15 +121,15 @@
         [StringLength(250)]
         public string MkzName { get; set; }
 
-        public int MkzCode1 { get; set; }
+        public long MkzCode1 { get; set; }
 
-        public int MkzCode2 { get; set; }
+        public long MkzCode2 { get; set; }
 
-        public int MkzCode3 { get; set; }
+        public long MkzCode3 { get; set; }
 
-        public int? MkzCode4 { get; set; }
+        public long? MkzCode4 { get; set; }
 
-        public int? MkzCode5 { get; set; }
+        public long? MkzCode5 { get; set; }
 
 
         [StringLength(50)]
